simulated-devices/pyproject.toml
[project]
name = "simulated-sensors"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
